New class to generate a combobox for the plate property selection
The class facilitates the creation of the combox for the plate property seleciton that is necessairy for the filter functionality, which in the future should provide the possibility to filter not only barcodes but any plate attribute.
At this point it is not quite clear which is the best way to configure the combobox, wether the filtering attributes should be taken from the plate class or if they should be configured via the node configuration dialog (the later is easier and more stable but requires additional user interaction).