indroducing new HeatMapModel2
The HeatMapModel has to change quite a bit in order to be able to handle arbitrary color gradients, so a new class was introduced to be able to let the old and the new GUI coexist for some more time.
Consequentially code was copied, the HeatMapModel usages refactored to HeatMapModel2 and the old classes are beeing tagged deprecated.