Plate sorting runs as it should (on the first glance, needs thorrow testing)
PlateAttributeDialog: changing the internal handling of the selection order. Inseatd of passing a hashmap to the heatmapmodel, it uses a TreeMap now, so the Attributes can be retrieved easily in the order they were selected.
Adding the possibility to sort in ascending or descending order.
PlateComparators: added the missing comparators.