Changing to the new name scheme
in commit abda90272c947123ca08f4d958e05bca027173d1 (introducing new HeatMapModel2) a hole refactoring process started...
This commit is still a consequence of that, changing references to the new copies of classes. The PlateViewer remains to be re-implemented.
I forgot to mention, that the new class names are beside allowing the coexistence of old and new gui, should reflect what they do, rather than what they are in the Java class Hierarchy.