commenting sortHeamaps method, removing constructor code redundancy.
the ScreenViewers ScreenHeatMaps panel has a sortHeatmaps method, that does not make any sense to me. The sorting functionality is provided by the HeatMapModel. So I commented it out to see if it is missing.
The constructors of the ScreenViewer looked messy. while removing redundant code, I found some stumbling blocks in the initialisation cascade and fixed them