Adding Trellis configuration functionality
The number of rows and columns of the trellis can now be calculated automatically or set manually.
Additionaly one has the option to fix the plate proportions. If the "Fix Plate Proportions" is checked, the plates won't be dimensioned to fill the panel, but keep their dimensions according to the current zoom level.
The way of how to determine the number of columns and rows also changed slightly, taking the gap between plates and plate borders into account.