Heatmap trellis modifications
Changing the border type.
Adding the code to trunkate the barcode if necessairy.
Setting the background color of the HeatScreen panel to the color corresponding to empty wells according to the ScreenColorScheme.
Fixing a minor bug in the zoom functionality. (The widht/height ratio was deteriorating because of the "rounding" operation, now the ratio is approximated each zoom action again.)