fixing a bug of the ReadoutRescaleStrategy functionality ("Outlier Handling" menu item)
After moving the color and toolbar from the HeatTrellis to the ScreenViewer, the color rendering of the heatmaps would not be triggered rightly. After switching the "Outlier Handling" method (ReadoutScaleStrategy), all the heatmaps would be gray.
While searching, I found that many methods of the HeatTrellis are redundant, so I commented a lot of stuff and put the crreateHeatMaps method in place. This takes the maps directly from the heatMapModel, and is only called via the ChangeListener. Before the HeatMaps were stored and synchronized, but they way it was done I am not sure If that made it more efficient, in my eyes it was most of all a source for bugs.
Finally I was adding a ChangeListener to the HeatMapColorToolBar. By updating the colorbar, the getMinValue and getMaxValue of the RescaleStrategy are called. Weirdly enough after calling the configure method of that class the getMinValue or getMaxvalue have to be called so the internal minMap and maxMap fields are updated. Those are necessairy so that the getReadoutColor method from the heatMapModel actually get a color value other than the one for missing values.