Adding single plate selection functionality to the HeatTrellis
implementing PlateSelectionController as a MouseListener.
Adding selection methods in the HeatMapModel.
Changing the constructor (hopefully the last time) of the HeatMapMenu. Also adapting the default for the selection marking as well as some access methods. Some renaming happened to avoid confusion with the filtering related methods.
Removing the conditions to select the background colors in the HeatScreen class. Since only entire plates are selected in the Trellis-View, the HeatScreen renderer doe not have to have a selection color gate for singel wells.