implementing drag selection on the trellis view
Instead of having a nested class, the HeatTrellis now implements the MouseChangeListener.
Removing the MouseListeners from the HeatScreen to avoid conflict (no more double clicks).
adding right click to open a PlateViewer.
Moved the code around a bit in the HeatTrellis class (I fear it's getting a bit confusing otherwise)