Synchronizing the selection between the ScreenViewer and the PlateViewer(s)
the instance of the HeatMapModel that collects the ChangeListeners lies in the HeatTrellis.
adding methods to update the ChangeListeners from the HeatMapModel.
moving the creation of the PlateViewer directly in the MouseListener method of the HeatTrellis (from the HeatScreen). This simplifies the updating of the ChangeListeners in the HeatMapModel (less accessors needed).