adding option for global color scaling between windows and proper window handling
The Trellis menu of the ScreenMenu now provides the option to us a global color scale. The alternative is that the trellis has a scale calculated over all plates (global) and the plateViewers calculate their own.
Since it might be confusing to have PlateViewers synchronizing with the ScreenViewer and others not, all the windows are closed when toggling the option.
 This lead to the implementation of a window handler in the HeatTrellis.