implementing HiLite functionality in the GUI
adding action listeners and actions in the menus. moved the hilite filtering to the trellis menu (since it does not concern the single plate).
adding necessairy fields in the HeatMapModel.
implemented a getPlatesToDisplay method. before this was in the Trellis, but with the hiliting it had to be adapted too and I found it belongs to the model.
adding conditions in the rederers (HeatWell, HeatTrellis) to pick the right color
changing the ScreenColorScheme to use the KNIME colors for selection and hiliting.