The NodeView does not eat JFrames!!!
This lead to the following series of modifications:
The ScreenViewer now extends JFrame. This entailed some porblems of how the menubar was handled before.  So instead of having tow menubars (one for the ScreenViewer and another for the PlateViewer) we have now simple objects for each menu, that can be easily added to any frame. On the way there the HeatMapViewer  interface was turned over a bit...