adding KNIME color support
it started with the new test class, providing an adequat dataset: PlateHeatMapViewerTest2.
to retrieve the color while parsing the plate to the HeatMapModel, the DataTableSpec is needed. This lead to the modification of the HeatMapViewerNodeModels method signature change.
The Well model needed a color field.
There were a lot of changes necessairy in the HeatMapModel. Since the knime colors are a parallel way of defining overlays, some overlay attribute gating needed to be added.
Minimal changes happened in the HeatMapInputToolbar to make the "KNIME Color Model" option available if row columns are defined in the input table.
Finally the OverlayLegend nedded to be adapted too. Here I encountered the problem, that up to now, there are no proper group names available (see new todos).