Finishing the HiLite functionality
Registering the HiLiteListener in the HeatMapViewerNodeView and initialization of HiLites registered at the moment of the view instantiation.
Migrating the HiLiteListener interface implementation from the ScreenViewer to the HeatMapModel2.
Adding INPORT field to the HeatMapViewerNodeModel that is now used to get the HiLiteHandler in der HeatMapModel2 and the HiLiteMenu.
Propagating the HiLiteHandler to the PlateViewer so that all the HeatMapViewer windows are synchronized.