Adding warning when the view saves it's configurations
When the node view of the HeatMapViewer closes, it saves the configurations. Since this might take a moment, the user could think that the program hangs. For information there is a warning plotted now. I also added a TODO, to implement a dialog in the serialization routine of the node model so it can't be overseen. The dialog should also give the option to cancel the routine.