Maintaining the order of readouts and factos as given by the input table (issue #8)
This factor and readout lists are now passed to the HeatMapModel as additional fields. I am not very happy with this solutions, in my opinion LinkedHashMaps to store the well statistics and annotations would have been better, to keep the order as derivable parameter. But I did not want to remodel the Well class (I could not find a neat way of running old and new in parallel without it being confusing).  This way there is another possible source of errors.
I also had to adapt the Tests to accomodate the additional functionality performed by the HeatMapViewerNodeModel.