fixing a small bug in the overlay popupmenu
Previously, when an empty string was added to the options list of the overlay popupmenu, the variable used to carry the options list was not seperated from the "annotations" field of the HeatMapModel. Now the list is a new instance, not modifying the content of the "annotations" field of the HeatMapModel anymore.