Adding the well annotations to the table of the WellViewer (issue #9)
... before only the readouts (well statistics) were displayed.
To get the readouts and annotations in the same order as in the configuration dialog of the node, I use the "readouts" and "annotations" fields of the HeatMapModel. To reach the data model from the WellViewer, the HeatWell, from which the WellViewer is triggered, had to provide a getter for the data model.
The Table of the WellViewer now has a tooltip for the annotation/readout values too. Before it was not necessary, having only numbers. The strings from the annotations however could exceed the column width and thus a tooltip revealing the entire string will be useful in some cases.
The readoutTable variable was refactored to wellTable.