deactivating the right click functionality of the ColorGradientSlider (issue #7)
the contextual menu evoked by the right click had some quirks; If the thumb was not hit exactly in the middle it would move a bit before the contextual menu (color menu) would open. Since the contextual menu does not offer the ability to cancel the selection, it's better to choose the colors with the dialog (evoked by double clicking on the thumb).