Bug fix for the error on closing the HeatMapViewerNodeView (issue #2)
I put a condition in the seralizeViewConfiguration method that checks if a binary file for the view configurations exists. If it exists, the node has been loaded already and the configuration has to be saved for the case that the workflow is not modified (bad hack). If the file does not exist, the routine is skipped when closing the view but will be taken care of in the saveInternals method of the node model, that will be called on closing knime, or the workflow, since there is a new node which sets the workflow modification flag.
It would be nicer to just set the workflow-is-modified (dirty) flag, so the view configuration has to be saved only once on closing the workflow and not each time on closing the view.