Remodeling the HeatMapViewerNodeDialog.
There were some problems with the node settings when using the AbstractScreenTrafoDialog to create a column selector acting on a list selector containing the groups available in the nominal column. So I added code to create the custom dialog components in this class itself and modified the createControls method accordingly.