Adding an image output port to the HeatMapViewer (feature request #19)
The configure and execute routines of the HeatMapViewerNodeModel and an additional method has been added that creates the image output.
The HeatTrellis and HeatMapColorToolBar needed some adjustmets to makes certain properties available from outside. For instance the font need to be bigger when rendering it for an image.
Finally the AbstractNodeModel needed a new constructor to be able to hold the new port configuration of the HeatMapViewerNodeModel