Added the hcscore sources
The hcscore was built with maven and then put into the lib directory of
the hcs-tools with a shell script.
I added the hcscore sources to the project with the intention, that this plugin
will be built together with the hcs-tools (now de.mpicbg.knime.hcs.base)
from within eclipse.